Consolidate GPU vendor support into CI wrapper
