+++ zstd/zstdgpu_ci_tests/zstdgpu_ci_tests.vcxproj
@@ -222,15 +222,12 @@
   <!-- Source files -->
   <ItemGroup>
     <ClCompile Include="adversarial_manifest.cpp" />
-    <ClCompile Include="gpu_vendor_args.cpp" />
-    <ClCompile Include="gpu_vendor_args_tests.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="zstd_frame_size.cpp" />
     <ClCompile Include="zstdgpu_ci_tests.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="adversarial_manifest.h" />
-    <ClInclude Include="gpu_vendor_args.h" />
     <ClInclude Include="zstd_frame_size.h" />
     <ClInclude Include="zstdgpu_ci_tests.h" />
   </ItemGroup>
